Move internal/testcomponents to service/internal/testcomponents, reduce scope (#6133)
