+++ service/internal/testcomponents/example_processor.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package testcomponents // import "go.opentelemetry.io/collector/internal/testcomponents"
+package testcomponents // import "go.opentelemetry.io/collector/service/internal/testcomponents"
 
 import (
 	"context"
