+++ service/internal/testcomponents/example_processor_test.go
