+++ service/internal/testcomponents/example_receiver_test.go
